+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/VirtualizedFileSystemTest.java
@@ -95,13 +95,23 @@ import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.TruffleLanguage.Env;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.RootNode;
+import java.nio.file.FileVisitOption;
+import java.nio.file.FileVisitResult;
+import java.nio.file.FileVisitor;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.nio.file.attribute.GroupPrincipal;
+import java.nio.file.attribute.UserPrincipal;
+import java.util.HashSet;
 
 @RunWith(Parameterized.class)
 public class VirtualizedFileSystemTest {
 
     private static final String LANGAUGE_ID = "virtualised-fs-lang";
     private static final String FOLDER_EXISTING = "folder";
+    private static final String FOLDER_EXISTING_INNER1 = "folder1";
+    private static final String FOLDER_EXISTING_INNER2 = "folder2";
     private static final String FILE_EXISTING = "existing_file.txt";
+    private static final String FILE_EXISTING2 = "existing_file2.txt";
     private static final String FILE_EXISTING_CONTENT = "Existing File Content";
     private static final String FILE_EXISTING_WRITE_MMAP = "write_mmap.txt";
     private static final String FILE_EXISTING_DELETE = "delete.txt";
@@ -111,6 +121,11 @@ public class VirtualizedFileSystemTest {
     private static final String FILE_NEW_CREATE_DIR = "new_dir";
     private static final String FILE_NEW_CREATE_FILE = "new_file.txt";
     private static final String FILE_NEW_RENAME = "new_rename.txt";
+    private static final String FILE_NEW_LINK = "new_link.txt";
+    private static final String FILE_NEW_SYMLINK = "new_symlink.txt";
+    private static final String FILE_NEW_COPY = "new_copy.txt";
+    private static final String FOLDER_NEW_COPY = "folder_copy";
+
     private static Collection<Configuration> cfgs;
     private static Consumer<Env> languageAction;
 
@@ -702,6 +717,440 @@ public class VirtualizedFileSystemTest {
         ctx.eval(LANGAUGE_ID, "");
     }
 
+    @Test
+    public void testRelativize() {
+        Context ctx = cfg.getContext();
+        languageAction = (Env env) -> {
+            TruffleFile parent = env.getTruffleFile("/test/parent");
+            TruffleFile child = env.getTruffleFile("/test/parent/child");
+            TruffleFile relative = parent.relativize(child);
+            Assert.assertEquals("child", relative.getPath());
+            Assert.assertEquals(child, parent.resolve(relative.getPath()));
+            child = env.getTruffleFile("/test/parent/child/inner");
+            relative = parent.relativize(child);
+            Assert.assertEquals("child/inner", relative.getPath());
+            Assert.assertEquals(child, parent.resolve(relative.getPath()));
+            TruffleFile sibling = env.getTruffleFile("/test/sibling");
+            relative = parent.relativize(sibling);
+            Assert.assertEquals("../sibling", relative.getPath());
+            Assert.assertEquals(sibling.normalize(), parent.resolve(relative.getPath()).normalize());
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
+    @Test
+    public void testStartsWith() {
+        Context ctx = cfg.getContext();
+        languageAction = (Env env) -> {
+            TruffleFile testAbsolute = env.getTruffleFile("/test");
+            TruffleFile testParentAbsolute = env.getTruffleFile("/test/parent");
+            TruffleFile testSiblingAbolute = env.getTruffleFile("/test/sibling");
+            TruffleFile testParentSiblingAbsolute = env.getTruffleFile("/test/parent/sibling");
+            TruffleFile teAbsolute = env.getTruffleFile("/te");
+            TruffleFile testParentChildAbsolute = env.getTruffleFile("/test/parent/child");
+            TruffleFile testRelative = env.getTruffleFile("test");
+            TruffleFile testParentRelative = env.getTruffleFile("test/parent");
+            TruffleFile testSiblingRelative = env.getTruffleFile("test/sibling");
+            TruffleFile testParentSiblingRelative = env.getTruffleFile("test/parent/sibling");
+            TruffleFile teRelative = env.getTruffleFile("te");
+            TruffleFile testParentChildRelative = env.getTruffleFile("test/parent/child");
+            Assert.assertTrue(testParentChildAbsolute.startsWith(testAbsolute));
+            Assert.assertTrue(testParentChildAbsolute.startsWith(testAbsolute.getPath()));
+            Assert.assertTrue(testParentChildAbsolute.startsWith(testParentAbsolute));
+            Assert.assertTrue(testParentChildAbsolute.startsWith(testParentAbsolute.getPath()));
+            Assert.assertTrue(testParentChildAbsolute.startsWith(testParentChildAbsolute));
+            Assert.assertTrue(testParentChildAbsolute.startsWith(testParentChildAbsolute.getPath()));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testSiblingAbolute));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testSiblingAbolute.getPath()));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testParentSiblingAbsolute));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testParentSiblingAbsolute.getPath()));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(teAbsolute));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(teAbsolute.getPath()));
+            Assert.assertTrue(testParentChildRelative.startsWith(testRelative));
+            Assert.assertTrue(testParentChildRelative.startsWith(testRelative.getPath()));
+            Assert.assertTrue(testParentChildRelative.startsWith(testParentRelative));
+            Assert.assertTrue(testParentChildRelative.startsWith(testParentRelative.getPath()));
+            Assert.assertTrue(testParentChildRelative.startsWith(testParentChildRelative));
+            Assert.assertTrue(testParentChildRelative.startsWith(testParentChildRelative.getPath()));
+            Assert.assertFalse(testParentChildRelative.startsWith(testSiblingRelative));
+            Assert.assertFalse(testParentChildRelative.startsWith(testSiblingRelative.getPath()));
+            Assert.assertFalse(testParentChildRelative.startsWith(testParentSiblingRelative));
+            Assert.assertFalse(testParentChildRelative.startsWith(testParentSiblingRelative.getPath()));
+            Assert.assertFalse(testParentChildRelative.startsWith(teRelative));
+            Assert.assertFalse(testParentChildRelative.startsWith(teRelative.getPath()));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testRelative));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testRelative.getPath()));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testParentRelative));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testParentRelative.getPath()));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testParentChildRelative));
+            Assert.assertFalse(testParentChildAbsolute.startsWith(testParentChildRelative.getPath()));
+            Assert.assertFalse(testParentChildRelative.startsWith(testAbsolute));
+            Assert.assertFalse(testParentChildRelative.startsWith(testAbsolute.getPath()));
+            Assert.assertFalse(testParentChildRelative.startsWith(testParentAbsolute));
+            Assert.assertFalse(testParentChildRelative.startsWith(testParentAbsolute.getPath()));
+            Assert.assertFalse(testParentChildRelative.startsWith(testParentChildAbsolute));
+            Assert.assertFalse(testParentChildRelative.startsWith(testParentChildAbsolute.getPath()));
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
+    @Test
+    public void testEndsWith() {
+        Context ctx = cfg.getContext();
+        languageAction = (Env env) -> {
+            TruffleFile testParentInnerAbsolute = env.getTruffleFile("/test/parent/inner");
+            TruffleFile testParentInnerRelative = env.getTruffleFile("test/parent/inner");
+            TruffleFile innerAbsolute = env.getTruffleFile("/inner");
+            TruffleFile innerRelative = env.getTruffleFile("inner");
+            TruffleFile parentInnerAbsolute = env.getTruffleFile("/parent/inner");
+            TruffleFile parentInnerRelative = env.getTruffleFile("parent/inner");
+            TruffleFile nnerRelative = env.getTruffleFile("nner");
+            TruffleFile testParentSiblingAbsolute = env.getTruffleFile("/test/parent/sibling");
+            TruffleFile testParentSiblingRelative = env.getTruffleFile("test/parent/sibling");
+            TruffleFile testParentInnerChildAbsolute = env.getTruffleFile("/test/parent/inner/child");
+            TruffleFile testParentInnerChildRelative = env.getTruffleFile("test/parent/inner/child");
+
+            Assert.assertTrue(testParentInnerAbsolute.endsWith(testParentInnerAbsolute));
+            Assert.assertTrue(testParentInnerAbsolute.endsWith(testParentInnerAbsolute.getPath()));
+            Assert.assertTrue(testParentInnerAbsolute.endsWith(testParentInnerRelative));
+            Assert.assertTrue(testParentInnerAbsolute.endsWith(testParentInnerRelative.getPath()));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(innerAbsolute));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(innerAbsolute.getPath()));
+            Assert.assertTrue(testParentInnerAbsolute.endsWith(innerRelative));
+            Assert.assertTrue(testParentInnerAbsolute.endsWith(innerRelative.getPath()));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(parentInnerAbsolute));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(parentInnerAbsolute.getPath()));
+            Assert.assertTrue(testParentInnerAbsolute.endsWith(parentInnerRelative));
+            Assert.assertTrue(testParentInnerAbsolute.endsWith(parentInnerRelative.getPath()));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(nnerRelative));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(nnerRelative.getPath()));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(testParentSiblingAbsolute));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(testParentSiblingAbsolute.getPath()));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(testParentSiblingRelative));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(testParentSiblingRelative.getPath()));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(testParentInnerChildAbsolute));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(testParentInnerChildAbsolute.getPath()));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(testParentInnerChildRelative));
+            Assert.assertFalse(testParentInnerAbsolute.endsWith(testParentInnerChildRelative.getPath()));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentInnerAbsolute));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentInnerAbsolute.getPath()));
+            Assert.assertTrue(testParentInnerRelative.endsWith(testParentInnerRelative));
+            Assert.assertTrue(testParentInnerRelative.endsWith(testParentInnerRelative.getPath()));
+            Assert.assertFalse(testParentInnerRelative.endsWith(innerAbsolute));
+            Assert.assertFalse(testParentInnerRelative.endsWith(innerAbsolute.getPath()));
+            Assert.assertTrue(testParentInnerRelative.endsWith(innerRelative));
+            Assert.assertTrue(testParentInnerRelative.endsWith(innerRelative.getPath()));
+            Assert.assertFalse(testParentInnerRelative.endsWith(parentInnerAbsolute));
+            Assert.assertFalse(testParentInnerRelative.endsWith(parentInnerAbsolute.getPath()));
+            Assert.assertTrue(testParentInnerRelative.endsWith(parentInnerRelative));
+            Assert.assertTrue(testParentInnerRelative.endsWith(parentInnerRelative.getPath()));
+            Assert.assertFalse(testParentInnerRelative.endsWith(nnerRelative));
+            Assert.assertFalse(testParentInnerRelative.endsWith(nnerRelative.getPath()));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentSiblingAbsolute));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentSiblingAbsolute.getPath()));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentSiblingRelative));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentSiblingRelative.getPath()));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentInnerChildAbsolute));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentInnerChildAbsolute.getPath()));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentInnerChildRelative));
+            Assert.assertFalse(testParentInnerRelative.endsWith(testParentInnerChildRelative.getPath()));
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
+    @Test
+    public void testNewDirectoryStream() {
+        Context ctx = cfg.getContext();
+        Path path = cfg.getPath();
+        boolean canRead = cfg.canRead();
+        languageAction = (Env env) -> {
+            Path folderExisting = path.resolve(FOLDER_EXISTING);
+            TruffleFile file = cfg.needsURI() ? env.getTruffleFile(folderExisting.toUri()) : env.getTruffleFile(folderExisting.toString());
+            Set<String> expected = new HashSet<>();
+            Collections.addAll(expected, FILE_EXISTING, FOLDER_EXISTING_INNER1, FOLDER_EXISTING_INNER2);
+            try (DirectoryStream<TruffleFile> stream = file.newDirectoryStream()) {
+                Assert.assertTrue(cfg.formatErrorMessage("Expected SecurityException"), canRead);
+                Set<String> result = new HashSet<>();
+                for (TruffleFile child : stream) {
+                    result.add(child.getName());
+                }
+                Assert.assertEquals(expected, result);
+            } catch (SecurityException se) {
+                Assert.assertFalse(cfg.formatErrorMessage("Unexpected SecurityException"), canRead);
+            } catch (IOException ioe) {
+                throw new AssertionError(cfg.formatErrorMessage(ioe.getMessage()), ioe);
+            }
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
+    @Test
+    public void testVisit() {
+        final Context ctx = cfg.getContext();
+        final Path path = cfg.getPath();
+        final boolean canRead = cfg.canRead();
+        languageAction = (Env env) -> {
+            TruffleFile root = env.getTruffleFile(path.toString());
+            TruffleFile existingFolder = root.resolve(FOLDER_EXISTING);
+            try {
+                // @formatter:off
+                TestVisitor visitor = TestVisitor.newBuilder(existingFolder).
+                        folder(FOLDER_EXISTING_INNER1).
+                            file(FILE_EXISTING).
+                            file(FILE_EXISTING2).
+                        end().
+                        folder(FOLDER_EXISTING_INNER2).
+                            folder(FOLDER_EXISTING_INNER1).
+                            end().
+                            folder(FOLDER_EXISTING_INNER2).
+                                file(FILE_EXISTING).
+                            end().
+                            file(FILE_EXISTING).
+                            file(FILE_EXISTING2).
+                        end().
+                        file(FILE_EXISTING).
+                        build();
+                // @formatter:on
+                existingFolder.visit(visitor, Integer.MAX_VALUE, FileVisitOption.FOLLOW_LINKS);
+                Assert.assertTrue(cfg.formatErrorMessage("Expected SecurityException"), canRead);
+                visitor.assertConsumed();
+                TruffleFile existingFile = existingFolder.resolve(FILE_EXISTING);
+                visitor = TestVisitor.newBuilder(existingFile).build();
+                existingFile.visit(visitor, Integer.MAX_VALUE, FileVisitOption.FOLLOW_LINKS);
+                visitor.assertConsumed();
+                // @formatter:off
+                visitor = TestVisitor.newBuilder(existingFolder).
+                        file(FOLDER_EXISTING_INNER1).
+                        file(FOLDER_EXISTING_INNER2).
+                        file(FILE_EXISTING).
+                        build();
+                // @formatter:on
+                existingFolder.visit(visitor, 1, FileVisitOption.FOLLOW_LINKS);
+                visitor.assertConsumed();
+                // @formatter:off
+                visitor = TestVisitor.newBuilder(existingFolder).
+                        folder(FOLDER_EXISTING_INNER1).
+                            file(FILE_EXISTING).
+                            file(FILE_EXISTING2).
+                        end().
+                        folder(FOLDER_EXISTING_INNER2).
+                            file(FOLDER_EXISTING_INNER1).
+                            file(FOLDER_EXISTING_INNER2).
+                            file(FILE_EXISTING).
+                            file(FILE_EXISTING2).
+                        end().
+                        file(FILE_EXISTING).
+                        build();
+                // @formatter:on
+                existingFolder.visit(visitor, 2, FileVisitOption.FOLLOW_LINKS);
+                visitor.assertConsumed();
+                // @formatter:off
+                visitor = TestVisitor.newBuilder(existingFolder).
+                        folder(FOLDER_EXISTING_INNER1).
+                            file(FILE_EXISTING).
+                            file(FILE_EXISTING2).
+                        end().
+                        folder(FOLDER_EXISTING_INNER2).
+                            skipSubTree().
+                        end().
+                        file(FILE_EXISTING).
+                        build();
+                // @formatter:on
+                existingFolder.visit(visitor, Integer.MAX_VALUE, FileVisitOption.FOLLOW_LINKS);
+                Assert.assertTrue(cfg.formatErrorMessage("Expected SecurityException"), canRead);
+                visitor.assertConsumed();
+                // TestVisitor cannot be used for SKIP_SIBLINGS due to random order of files on file
+                // system
+                FileVisitor<TruffleFile> fileVisitor = new FileVisitor<TruffleFile>() {
+
+                    private boolean skipReturned;
+                    private Set<TruffleFile> importantFiles;
+
+                    {
+                        importantFiles = new HashSet<>();
+                        importantFiles.add(existingFolder.resolve(FOLDER_EXISTING_INNER1));
+                        importantFiles.add(existingFolder.resolve(FOLDER_EXISTING_INNER2));
+                        importantFiles.add(existingFile);
+                    }
+
+                    @Override
+                    public FileVisitResult preVisitDirectory(TruffleFile dir, BasicFileAttributes attrs) throws IOException {
+                        return check(dir);
+                    }
+
+                    @Override
+                    public FileVisitResult visitFile(TruffleFile file, BasicFileAttributes attrs) throws IOException {
+                        return check(file);
+                    }
+
+                    @Override
+                    public FileVisitResult visitFileFailed(TruffleFile file, IOException exc) throws IOException {
+                        return FileVisitResult.TERMINATE;
+                    }
+
+                    @Override
+                    public FileVisitResult postVisitDirectory(TruffleFile dir, IOException exc) throws IOException {
+                        return FileVisitResult.CONTINUE;
+                    }
+
+                    private FileVisitResult check(TruffleFile file) {
+                        if (importantFiles.contains(file)) {
+                            if (skipReturned) {
+                                throw new AssertionError("Visited skipped sibling: " + file);
+                            } else {
+                                skipReturned = true;
+                                return FileVisitResult.SKIP_SIBLINGS;
+                            }
+                        }
+                        return FileVisitResult.CONTINUE;
+                    }
+                };
+                existingFolder.visit(fileVisitor, Integer.MAX_VALUE, FileVisitOption.FOLLOW_LINKS);
+            } catch (SecurityException se) {
+                Assert.assertFalse(cfg.formatErrorMessage("Unexpected SecurityException"), canRead);
+            } catch (IOException ioe) {
+                throw new AssertionError(cfg.formatErrorMessage(ioe.getMessage()), ioe);
+            }
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
+    @Test
+    public void testCreateLink() {
+        Context ctx = cfg.getContext();
+        Path path = cfg.getPath();
+        boolean canWrite = cfg.canWrite();
+        languageAction = (Env env) -> {
+            final TruffleFile root = cfg.needsURI() ? env.getTruffleFile(path.toUri()) : env.getTruffleFile(path.toString());
+            try {
+                TruffleFile target = root.resolve(FOLDER_EXISTING).resolve(FILE_EXISTING);
+                TruffleFile link = root.resolve(FILE_NEW_LINK);
+                link.createLink(target);
+                Assert.assertTrue(cfg.formatErrorMessage("Expected SecurityException"), canWrite);
+            } catch (SecurityException se) {
+                Assert.assertFalse(cfg.formatErrorMessage("Unexpected SecurityException"), canWrite);
+            } catch (IOException ioe) {
+                throw new AssertionError(cfg.formatErrorMessage(ioe.getMessage()), ioe);
+            } catch (UnsupportedOperationException uoe) {
+                // Links may not be supported on file system
+            }
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
+    @Test
+    public void testCreateSymbolicLink() {
+        Context ctx = cfg.getContext();
+        Path path = cfg.getPath();
+        boolean canWrite = cfg.canWrite();
+        languageAction = (Env env) -> {
+            TruffleFile root = cfg.needsURI() ? env.getTruffleFile(path.toUri()) : env.getTruffleFile(path.toString());
+            try {
+                TruffleFile target = root.resolve(FOLDER_EXISTING).resolve(FILE_EXISTING);
+                TruffleFile link = root.resolve(FILE_NEW_SYMLINK);
+                link.createSymbolicLink(target);
+                Assert.assertTrue(cfg.formatErrorMessage("Expected SecurityException"), canWrite);
+            } catch (SecurityException se) {
+                Assert.assertFalse(cfg.formatErrorMessage("Unexpected SecurityException"), canWrite);
+            } catch (IOException ioe) {
+                throw new AssertionError(cfg.formatErrorMessage(ioe.getMessage()), ioe);
+            } catch (UnsupportedOperationException uoe) {
+                // Symbolik links may not be supported on file system
+            }
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
+    @Test
+    public void testGetOwner() {
+        Context ctx = cfg.getContext();
+        Path path = cfg.getPath();
+        boolean canRead = cfg.canRead();
+        languageAction = (Env env) -> {
+            final TruffleFile root = cfg.needsURI() ? env.getTruffleFile(path.toUri()) : env.getTruffleFile(path.toString());
+            try {
+                TruffleFile file = root.resolve(FOLDER_EXISTING);
+                UserPrincipal owner = file.getOwner();
+                Assert.assertTrue(cfg.formatErrorMessage("Expected SecurityException"), canRead);
+                Assert.assertNotNull(owner);
+            } catch (SecurityException se) {
+                Assert.assertFalse(cfg.formatErrorMessage("Unexpected SecurityException"), canRead);
+            } catch (IOException ioe) {
+                throw new AssertionError(cfg.formatErrorMessage(ioe.getMessage()), ioe);
+            } catch (UnsupportedOperationException uoe) {
+                // Onwer may not be supported on file system
+            }
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
+    @Test
+    public void testGetGroup() {
+        Context ctx = cfg.getContext();
+        Path path = cfg.getPath();
+        boolean canRead = cfg.canRead();
+        languageAction = (Env env) -> {
+            final TruffleFile root = cfg.needsURI() ? env.getTruffleFile(path.toUri()) : env.getTruffleFile(path.toString());
+            try {
+                TruffleFile file = root.resolve(FOLDER_EXISTING);
+                GroupPrincipal group = file.getGroup();
+                Assert.assertTrue(cfg.formatErrorMessage("Expected SecurityException"), canRead);
+                Assert.assertNotNull(group);
+            } catch (SecurityException se) {
+                Assert.assertFalse(cfg.formatErrorMessage("Unexpected SecurityException"), canRead);
+            } catch (IOException ioe) {
+                throw new AssertionError(cfg.formatErrorMessage(ioe.getMessage()), ioe);
+            } catch (UnsupportedOperationException uoe) {
+                // Group may not be supported on file system
+            }
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
+    @Test
+    public void testCopy() {
+        Context ctx = cfg.getContext();
+        Path path = cfg.getPath();
+        boolean canRead = cfg.canRead();
+        boolean canWrite = cfg.canWrite();
+        languageAction = (Env env) -> {
+            TruffleFile root = cfg.needsURI() ? env.getTruffleFile(path.toUri()) : env.getTruffleFile(path.toString());
+            try {
+                TruffleFile file = root.resolve(FOLDER_EXISTING).resolve(FILE_EXISTING);
+                TruffleFile target = root.resolve(FILE_NEW_COPY);
+                file.copy(target);
+                Assert.assertTrue(cfg.formatErrorMessage("Expected SecurityException"), canWrite);
+                if (canRead) {
+                    Collection<? extends TruffleFile> children = root.list();
+                    boolean hasTarget = children.stream().filter((TruffleFile truffleFile) -> FILE_NEW_COPY.equals(truffleFile.getName())).findAny().isPresent();
+                    Assert.assertTrue(cfg.formatErrorMessage("Copied target file should exist"), hasTarget);
+                }
+            } catch (SecurityException se) {
+                Assert.assertFalse(cfg.formatErrorMessage("Unexpected SecurityException"), canWrite);
+            } catch (IOException ioe) {
+                throw new AssertionError(cfg.formatErrorMessage(ioe.getMessage()), ioe);
+            }
+            try {
+                TruffleFile folder = root.resolve(FOLDER_EXISTING);
+                TruffleFile target = root.resolve(FOLDER_NEW_COPY);
+                folder.copy(target);
+                Assert.assertTrue(cfg.formatErrorMessage("Expected SecurityException"), canWrite);
+                if (canRead) {
+                    Collection<? extends TruffleFile> children = root.list();
+                    boolean hasTarget = children.stream().filter((TruffleFile truffleFile) -> FOLDER_NEW_COPY.equals(truffleFile.getName())).findAny().isPresent();
+                    boolean hasChildren = !target.list().isEmpty();
+                    Assert.assertTrue(cfg.formatErrorMessage("Copied target file should exist"), hasTarget);
+                    Assert.assertTrue(cfg.formatErrorMessage("Copied target should not have children"), !hasChildren);
+                }
+            } catch (SecurityException se) {
+                Assert.assertFalse(cfg.formatErrorMessage("Unexpected SecurityException"), canWrite);
+            } catch (IOException ioe) {
+                throw new AssertionError(cfg.formatErrorMessage(ioe.getMessage()), ioe);
+            }
+        };
+        ctx.eval(LANGAUGE_ID, "");
+    }
+
     public static final class Configuration implements Closeable {
         private final String name;
         private final Context ctx;
@@ -866,8 +1315,17 @@ public class VirtualizedFileSystemTest {
     private static Path createContent(
                     final Path folder,
                     final FileSystem fs) throws IOException {
-        mkdirs(folder.resolve(FOLDER_EXISTING), fs);
-        write(folder.resolve(FOLDER_EXISTING).resolve(FILE_EXISTING), FILE_EXISTING_CONTENT.getBytes(StandardCharsets.UTF_8), fs);
+        Path existing = mkdirs(folder.resolve(FOLDER_EXISTING), fs);
+        Path l2i1 = mkdirs(existing.resolve(FOLDER_EXISTING_INNER1), fs);
+        write(l2i1.resolve(FILE_EXISTING), new byte[0], fs);
+        write(l2i1.resolve(FILE_EXISTING2), new byte[0], fs);
+        Path l2i2 = mkdirs(existing.resolve(FOLDER_EXISTING_INNER2), fs);
+        write(l2i2.resolve(FILE_EXISTING), new byte[0], fs);
+        write(l2i2.resolve(FILE_EXISTING2), new byte[0], fs);
+        mkdirs(l2i2.resolve(FOLDER_EXISTING_INNER1), fs);
+        Path l3i2 = mkdirs(l2i2.resolve(FOLDER_EXISTING_INNER2), fs);
+        write(l3i2.resolve(FILE_EXISTING), new byte[0], fs);
+        write(existing.resolve(FILE_EXISTING), FILE_EXISTING_CONTENT.getBytes(StandardCharsets.UTF_8), fs);
         touch(folder.resolve(FILE_EXISTING_WRITE_MMAP), fs);
         touch(folder.resolve(FILE_EXISTING_DELETE), fs);
         touch(folder.resolve(FILE_EXISTING_RENAME), fs);
@@ -1211,6 +1669,138 @@ public class VirtualizedFileSystemTest {
                 }
             }
             return null;
+        }
+    }
+
+    private static final class TestVisitor implements FileVisitor<TruffleFile> {
+
+        private Node current;
+
+        private TestVisitor(Node root) {
+            this.current = root;
+        }
+
+        void assertConsumed() {
+            Assert.assertTrue(current.children.isEmpty());
+        }
+
+        @Override
+        public FileVisitResult preVisitDirectory(TruffleFile dir, BasicFileAttributes attrs) throws IOException {
+            Node node = null;
+            for (Node child : current.children) {
+                if (child.file.equals(dir)) {
+                    node = child;
+                    break;
+                }
+            }
+            Assert.assertNotNull(node);
+            Assert.assertTrue(node.folder);
+            if (node.action == FileVisitResult.SKIP_SIBLINGS || node.action == FileVisitResult.SKIP_SUBTREE) {
+                current.children.remove(node);
+            } else {
+                current = node;
+            }
+            return node.action;
+        }
+
+        @Override
+        public FileVisitResult visitFile(TruffleFile file, BasicFileAttributes attrs) throws IOException {
+            Node node = null;
+            for (Node child : current.children) {
+                if (child.file.equals(file)) {
+                    node = child;
+                    break;
+                }
+            }
+            Assert.assertNotNull(node);
+            Assert.assertFalse(node.folder);
+            current.children.remove(node);
+            return node.action;
+        }
+
+        @Override
+        public FileVisitResult visitFileFailed(TruffleFile file, IOException exc) throws IOException {
+            return FileVisitResult.TERMINATE;
+        }
+
+        @Override
+        public FileVisitResult postVisitDirectory(TruffleFile dir, IOException exc) throws IOException {
+            Assert.assertTrue(current.children.isEmpty());
+            Node prev = current;
+            current = current.parent;
+            current.children.remove(prev);
+            return FileVisitResult.CONTINUE;
+        }
+
+        static Builder newBuilder(TruffleFile root) {
+            return new Builder(root);
+        }
+
+        static final class Builder {
+            private final Node rootNode;
+            private Node currentScope;
+
+            Builder(TruffleFile start) {
+                rootNode = new Node(null, (TruffleFile) null, true);
+                this.currentScope = new Node(rootNode, start, start.isDirectory());
+                rootNode.children.add(this.currentScope);
+            }
+
+            Builder folder(String name) {
+                Node newScope = new Node(currentScope, name, true);
+                currentScope.children.add(newScope);
+                currentScope = newScope;
+                return this;
+            }
+
+            Builder end() {
+                currentScope = currentScope.parent;
+                if (currentScope == null) {
+                    throw new IllegalStateException("Closing non opened folder.");
+                }
+                return this;
+            }
+
+            Builder file(String name) {
+                Node node = new Node(currentScope, name, false);
+                currentScope.children.add(node);
+                return this;
+            }
+
+            Builder skipSubTree() {
+                currentScope.action = FileVisitResult.SKIP_SUBTREE;
+                return this;
+            }
+
+            TestVisitor build() {
+                return new TestVisitor(rootNode);
+            }
+        }
+
+        private static final class Node {
+
+            private final Node parent;
+            private final TruffleFile file;
+            private final boolean folder;
+            private final Collection<Node> children;
+            private FileVisitResult action;
+
+            Node(Node parent, TruffleFile file, boolean folder) {
+                this.parent = parent;
+                this.file = file;
+                this.folder = folder;
+                this.children = folder ? new ArrayList<>() : Collections.emptyList();
+                this.action = FileVisitResult.CONTINUE;
+            }
+
+            Node(Node parent, String name, boolean folder) {
+                this(parent, parent.file.resolve(name), folder);
+            }
+
+            @Override
+            public String toString() {
+                return file.toString();
+            }
         }
     }
 }