+++ truffle/CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Added `SuspendedEvent.setReturnValue` to change the return value of the currently executed source location.
 * Deprecated `FrameSlot#getIndex` without replacement.
 * Added `TruffleInstrument.Env.startServer()` to get a virtual message-based server provided via `MessageTransport` service.
+* Added `TruffleFile.relativize`, `TruffleFile.startsWith`, `TruffleFile.endsWith`, `TruffleFile.createLink`,  `TruffleFile.createSymbolicLink`, `TruffleFile.getOwner`, `TruffleFile.getGroup`, `TruffleFile.newDirectoryStream`, `TruffleFile.visit`, `TruffleFile.copy` methods.
 
 ## Version 1.0.0 RC7
 * Truffle was relicensed from GPLv2 with CPE to Universal Permissive License (UPL).